requirements.txt
fastapi
uvicorn
python-dotenv
yt-dlp
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
ffmpeg-python
pydantic
typing-extensions